// src/models/enums/strings/addEventString.ts
export enum AddEventString {
    addNew = "Dodaj informacje o zdarzeniu",
    type = "Rodzaj zdarzenia",
    checkDate = "Wybierz datę zdarzenia",
    description = "Opis",
    descriptionLabel = `Spróbuj przekazać szczegółowe informacje, umozliwi to innym
    zaplanowanie bezpiecznej wycieczki.`,
    save = "Zapisz",
    cancel = "Anuluj",
    lookup = "Podgląd"

}